toolkit/components/url-classifier/tests/mochitest/test_cachemiss.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Bug 1272239 - Test gethash.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="classifierHelper.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>

<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">

<script src="head.js"></script>
<script class="testbody" type="text/javascript">
const MALWARE_LIST = "test-malware-simple";
const MALWARE_HOST = "malware.example.com/";

const UNWANTED_LIST = "test-unwanted-simple";
const UNWANTED_HOST = "unwanted.example.com/";

const GETHASH_URL = "http://mochi.test:8888/tests/toolkit/components/url-classifier/tests/mochitest/cache.sjs";

var shouldLoad = false;

var gPreGethashCounter = 0;
var gCurGethashCounter = 0;

function loadTestFrame() {
  return new Promise(function(resolve, reject) {
    var iframe = document.createElement("iframe");
    iframe.setAttribute("src", "gethashFrame.html");
    document.body.appendChild(iframe);

    iframe.onload = function() {
      document.body.removeChild(iframe);
      resolve();
    };
  }).then(getGethashCounter);
}

function getGethashCounter() {
  return new Promise(function(resolve, reject) {
    var xhr = new XMLHttpRequest;
    xhr.open("PUT", GETHASH_URL + "?gethashcount");
    xhr.setRequestHeader("Content-Type", "text/plain");
    xhr.onreadystatechange = function() {
      if (this.readyState == this.DONE) {
        gPreGethashCounter = gCurGethashCounter;
        gCurGethashCounter = parseInt(xhr.response);
        resolve();
      }
    };
    xhr.send();
  });
}

// add 4-bytes prefixes to local database, so when we access the url,
// it will trigger gethash request.
function addPrefixToDB(list, url) {
  var testData = [{ db: list, url, len: 4 }];

  return classifierHelper.addUrlToDB(testData)
    .catch(function(err) {
      ok(false, "Couldn't update classifier. Error code: " + err);
      // Abort test.
      SimpleTest.finish();
    });
}

// manually reset DB to make sure next test won't be affected by cache.
function reset() {
  return classifierHelper.resetDatabase();
}

// This test has to come before testPositiveCache to ensure gethash server doesn't
// contain completions.
function testNegativeCache() {
  SpecialPowers.DOMWindowUtils.clearSharedStyleSheetCache();
  shouldLoad = true;

  async function setup() {
    await classifierHelper.allowCompletion([MALWARE_LIST, UNWANTED_LIST], GETHASH_URL);

    // Only add prefix to database. not server, so gethash will not return
    // result.
    return Promise.all([
      addPrefixToDB(MALWARE_LIST, MALWARE_HOST),
      addPrefixToDB(UNWANTED_LIST, UNWANTED_HOST),
    ]);
  }

  return Promise.resolve()
    .then(setup)
    .then(() => loadTestFrame())
    .then(() => {
       ok(gCurGethashCounter > gPreGethashCounter, "Gethash request is triggered.");
})
    // Second load should not trigger gethash request because cache.
    .then(() => loadTestFrame())
    .then(() => {
       ok(gCurGethashCounter == gPreGethashCounter, "Gethash request is nottriggered.");
})
    .then(reset);
}

function testPositiveCache() {
  SpecialPowers.DOMWindowUtils.clearSharedStyleSheetCache();
  shouldLoad = false;

  async function setup() {
    await classifierHelper.allowCompletion([MALWARE_LIST, UNWANTED_LIST], GETHASH_URL);

    return Promise.all([
      addPrefixToDB(MALWARE_LIST, MALWARE_HOST),
      addPrefixToDB(UNWANTED_LIST, UNWANTED_HOST),
      addCompletionToServer(MALWARE_LIST, MALWARE_HOST, GETHASH_URL),
      addCompletionToServer(UNWANTED_LIST, UNWANTED_HOST, GETHASH_URL),
    ]);
  }

  return Promise.resolve()
    .then(setup)
    .then(() => loadTestFrame())
    .then(() => {
       ok(gCurGethashCounter > gPreGethashCounter, "Gethash request is triggered.");
})
    // Second load should not trigger gethash request because cache.
    .then(() => loadTestFrame())
    .then(() => {
       ok(gCurGethashCounter == gPreGethashCounter, "Gethash request is nottriggered.");
})
    .then(reset);
}

function runTest() {
  Promise.resolve()
    // This test resources get blocked when gethash returns successfully
    .then(classifierHelper.waitForInit)
    .then(testNegativeCache)
    .then(testPositiveCache)
    .then(function() {
      SimpleTest.finish();
    }).catch(function(e) {
      ok(false, "Some test failed with error " + e);
      SimpleTest.finish();
    });
}

SimpleTest.waitForExplicitFinish();

// 'network.predictor.enabled' is disabled because if other testcase load
// evil.js, evil.css ...etc resources, it may cause we load them from cache
// directly and bypass classifier check
SpecialPowers.pushPrefEnv({"set": [
  ["browser.safebrowsing.malware.enabled", true],
  ["urlclassifier.malwareTable", "test-malware-simple,test-unwanted-simple"],
  ["network.predictor.enabled", false],
  ["urlclassifier.gethash.timeout_ms", 30000],
]}, runTest);

</script>
</pre>
</body>
</html>
